ADD :: user feature 리팩토링
